Added in Achievement management class. Added local high score saving. Added ability to view leader board from in game menu.

+++ Paint/branches/PaintRunner - 2/PaintRunner/GameCenter/GCHelper.h
@@ -27,6 +27,6 @@
         achievementsToReport:(NSMutableArray *)achievementsToReport;
 - (void)reportAchievement:(NSString *)identifier 
           percentComplete:(double)percentComplete;
-- (void)reportScore:(NSString *)identifier score:(int)score;
+- (void)reportScore:(NSString *)identifier score:(double)score;
 
 @end
